update menu and read button status
