+++ config/__init__.py
@@ -19,6 +19,7 @@ DEVICE_ATS = device_config['ats']
 CLIENT = mqtt.TBGatewayMqttClient(host=HOST, port=10883, token=ACCESS_TOKEN)
 IO_PORT = '/dev/ttyS0'
 # IO_PORT = 'COM3'
+# IO_PORT = 'COM4'
 BAUDRATE = 115200
 # uncomment when test auto update firmware
 UPDATE_PERIOD = 600
@@ -36,6 +37,8 @@ cmd_led = {}
 cmd_lcd = {}
 cmd_sa = {}
 dict_cmd = {}
+
+button_status = [0]
 
 update_attributes_lock = threading.Lock()
 telemetries_lock = threading.Lock()
